# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GDGOC Hacktoberfest Backend",
    "description": "Event-management API for GDGOC Hacktoberfest open-source contribution tracking.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": [
          "Auth & Verification"
        ],
        "summary": "Register as a new student",
        "description": "Register student with official PSIT roll number and name.\nAccount is created in unverified state pending ID card / QR verification.",
        "operationId": "signup_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth & Verification"
        ],
        "summary": "Log in with roll number or email",
        "description": "Authenticate student, maintainer, or admin.\nIssues a cryptographically signed JWT scoped by user ID and role.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth & Verification"
        ],
        "summary": "Get authenticated user session",
        "description": "Retrieve profile details for the currently logged-in user session.\nProtected by JWT Bearer token authentication.",
        "operationId": "get_me_auth_me_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-id": {
      "post": {
        "tags": [
          "Auth & Verification"
        ],
        "summary": "Submit ID card and QR code for verification",
        "description": "Server-side student verification:\n1. Validates ID card photo format and size.\n2. Performs server-side cross-check of roll number and fuzzy name against portal record.\n3. Auto-verifies on exact roll + fuzzy name match; routes to admin review queue on mismatch/missing data.",
        "operationId": "verify_id_card_auth_verify_id_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyIDCardRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyIDCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/pending-verifications": {
      "get": {
        "tags": [
          "Auth & Verification"
        ],
        "summary": "Admin review queue for pending verifications",
        "description": "Admin-only review queue: returns unverified students who uploaded ID card photos\nthat require manual review and verification decision.",
        "operationId": "get_pending_verifications_auth_pending_verifications_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManualReviewItemResponse"
                  },
                  "title": "Response Get Pending Verifications Auth Pending Verifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-manual/{student_id}": {
      "post": {
        "tags": [
          "Auth & Verification"
        ],
        "summary": "Admin approve or reject student verification",
        "description": "Admin decision: approve or reject student verification.\nIf approved, student's account is verified with method 'manual'.",
        "operationId": "manual_verify_student_auth_verify_manual__student_id__post",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Student Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualReviewActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/github/link": {
      "post": {
        "tags": [
          "Auth & Verification"
        ],
        "summary": "Link GitHub account to student profile",
        "description": "Links authenticated student's profile to their verified GitHub username and GitHub ID.\nEnforces uniqueness across the event.",
        "operationId": "link_github_auth_github_link_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubOAuthLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubOAuthLinkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GitHubOAuthLinkRequest": {
        "properties": {
          "github_username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Github Username",
            "description": "GitHub username"
          },
          "github_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Id",
            "description": "GitHub numeric user ID"
          }
        },
        "type": "object",
        "required": [
          "github_username"
        ],
        "title": "GitHubOAuthLinkRequest"
      },
      "GitHubOAuthLinkResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "github_username": {
            "type": "string",
            "title": "Github Username"
          },
          "github_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "github_username",
          "message"
        ],
        "title": "GitHubOAuthLinkResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier",
            "description": "PSIT roll number or email address"
          }
        },
        "type": "object",
        "required": [
          "identifier"
        ],
        "title": "LoginRequest"
      },
...
      "ManualReviewActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "pattern": "^(approve|reject)$",
            "title": "Action",
            "description": "'approve' or 'reject'"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Optional admin review note"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ManualReviewActionRequest"
      },
      "ManualReviewItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "id_card_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Card Image Url"
          },
          "qr_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Token"
          },
          "portal_snapshot_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portal Snapshot Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "psit_roll_no",
          "verified",
          "created_at"
        ],
        "title": "ManualReviewItemResponse"
      },
...
      "SignupRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 2,
            "title": "Name",
            "description": "Full student name as on PSIT ID card"
          },
          "email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 5,
            "title": "Email",
            "description": "Student email address"
          },
          "psit_roll_no": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Psit Roll No",
            "description": "Official PSIT roll number"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "psit_roll_no"
        ],
        "title": "SignupRequest"
      },
      "SignupResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "pending_verification"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "psit_roll_no",
          "role",
          "verified",
          "created_at"
        ],
        "title": "SignupResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfileResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse"
      },
...
      "UserProfileResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "verification_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          },
          "github_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "psit_roll_no",
          "verified",
          "role",
          "created_at",
          "updated_at"
        ],
        "title": "UserProfileResponse"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "student",
          "maintainer",
          "admin"
        ],
        "title": "UserRole"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationMethod": {
        "type": "string",
        "enum": [
          "qr_auto",
          "manual"
        ],
        "title": "VerificationMethod"
      },
      "VerifyIDCardRequest": {
        "properties": {
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No",
            "description": "Student roll number to verify"
          },
          "qr_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qr Token",
            "description": "Decoded QR payload from ID card"
          },
          "id_card_image_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Card Image Base64",
            "description": "Base64-encoded ID card photo"
          },
          "portal_snapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portal Snapshot",
            "description": "PSIT portal student record"
          }
        },
        "type": "object",
        "required": [
          "psit_roll_no"
        ],
        "title": "VerifyIDCardRequest"
      },
      "VerifyIDCardResponse": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "verification_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "verified",
          "status",
          "message"
        ],
        "title": "VerifyIDCardResponse"
      },
...
    }
  }
}